Put Release Notes between Unreleased and Previous Version Heading
Instead of looking for a particular HTML block between the Unreleased heading and the previous version heading, we now put the release notes before the previous version heading.

Any possible content between Unreleased and the next Release Heading remains. (Maybe a note that this changelog is keept up to date automatically)